backend/requirements.txt
mysql-connector-python
python-dotenv
pandas
matplotlib
langchain
langgraph
langchain-community
langchain-experimental
langchain-core
langchain-openai
fastapi
uvicorn
pydantic